@@ -9,7 +9,6 @@ Bundler.setup
 $:.unshift(File.join(File.dirname(__FILE__), "lib"))
 require 'rust_radio'
 
-# bootstrap DataMapper
 config = YAML.load_file("config/config.yml")["config"]
 database = config["database"]
 username = database["username"]
@@ -18,10 +17,9 @@ hostname = database["hostname"]
 database = database["database"]
 
 ActiveRecord::Base.establish_connection(
-  :adapter  => "postgresql",
-  :host     => hostname,
-  :username => username,
-  :password => password,
-  :database => database
+  adapter:  "postgresql",
+  host:     hostname,
+  username: username,
+  password: password,
+  database: database
 )
-
