@@ -1,7 +1,7 @@
 module RustRadio
   class Show < ActiveRecord::Base
 
-    has_many :songs, :order => "sort_order"
+    has_many :songs, -> { order("sort_order ASC") }
 
     validates_uniqueness_of :folder_path
 
